Move mem2mem out of GDMA module
